backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-docx==1.1.0
pdfplumber==0.11.0
pylatexenc==2.10
python-multipart==0.0.9
Pillow==10.4.0
lxml==5.2.0
